Expiring _csrf cookie programmatically before fetching new token (#57)
* Expiring _csrf cookie programmatically before fetching new token

* Added attempt limit

+++ src/views/login/Login.tsx
@@ -25,9 +25,24 @@ const Login: React.FC = () => {
       setShowError(false);
     }
   }, [errorCode]);
+
+  const csrfTokenDeleted = (attemptCount = 0): boolean => {
+    if (attemptCount >= 10) {
+      return false;
+    }
+    if (
+      localStorageService.getCSRFToken() ||
+      document.cookie.includes('_csrf')
+    ) {
+      localStorageService.removeCSRFToken();
+      document.cookie = '_csrf=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT';
+      return csrfTokenDeleted(attemptCount + 1);
+    }
+    return true;
+  };
+
   useEffect(() => {
-    localStorageService.removeCSRFToken();
-    if (!localStorageService.getCSRFToken()) {
+    if (csrfTokenDeleted()) {
       dispatch(fetchCSRFToken());
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
